// apps/web-svelte/src/lib/api/generated/model/acquisitionDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AcquisitionSummary } from './acquisitionSummary';
import type { ReleaseCandidateView } from './releaseCandidateView';

export interface AcquisitionDetail {
  summary: AcquisitionSummary;
  candidates: ReleaseCandidateView[];
}